pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rfp-management-pipeline"
version = "0.1.0"
description = "AI-powered RFP analysis and proposal generation pipeline"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langchain==0.1.0",
    "langchain-google-genai==1.0.1",
    "langchain-community==0.0.10",
    "langchain-core==0.1.10",
    "transformers==4.35.2",
    "torch==2.1.1",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "numpy==1.24.3",
    "pymupdf==1.23.7"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "isort",
    "flake8"
]

[project.scripts]
rfp-pipeline = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"